requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
discord==2.3.2
discord.py==2.3.2
docopt==0.6.2
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.3.2
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
idna==3.4
ipykernel==6.25.1
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.1
msgpack==1.0.7
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.3
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==3.10.0
playsound==1.3.0
pooch==1.8.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pygame==2.4.0
Pygments==2.16.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.1.1
Pypubsub==4.0.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-vlc==3.0.18122
pytube==15.0.0
pywin32==306
pywin32-ctypes==0.2.1
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.6
yarg==0.1.9
yarl==1.9.4
